+++ express-joi-validation.js
@@ -59,7 +59,13 @@ function buildErrorString(err, container) {
   return ret
 }
 
-module.exports = function generateJoiMiddlewareInstance(cfg) {
+module.exports = function() {
+  throw new Error(
+    'express-joi-validation: exported member is no longer a factory function. use exported createValidator function instead'
+  )
+}
+
+module.exports.createValidator = function generateJoiMiddlewareInstance(cfg) {
   cfg = cfg || {} // default to an empty config
 
   const Joi = cfg.joi || require('@hapi/joi')